Fix warnings: Use correct printf format codes
This removes some warnings which are hiding warnings related to
64-bit architecure problems.

+++ interface/vchiq_arm/vchiq_test.h
@@ -77,8 +77,8 @@ struct test_params
       if (reason != _r) { \
          vcos_log_error("%d: expected callback reason " #_r ", got %d", __LINE__, reason); VCOS_BKPT; goto error_exit; \
       } \
-      else if ((int)VCHIQ_GET_SERVICE_USERDATA(service) != _u) { \
-         vcos_log_error("%d: expected userdata %d, got %d", __LINE__, _u, (int)VCHIQ_GET_SERVICE_USERDATA(service)); VCOS_BKPT; goto error_exit; \
+      else if ((intptr_t)VCHIQ_GET_SERVICE_USERDATA(service) != _u) { \
+         vcos_log_error("%d: expected userdata %d, got %zd", __LINE__, _u, (intptr_t)VCHIQ_GET_SERVICE_USERDATA(service)); VCOS_BKPT; goto error_exit; \
       } \
       else \
       { \
@@ -90,11 +90,11 @@ struct test_params
       if (reason != _r) { \
          vcos_log_error("%d: expected callback reason " #_r ", got %d", __LINE__, reason); VCOS_BKPT; goto error_exit; \
       } \
-      else if ((int)VCHIQ_GET_SERVICE_USERDATA(service) != _u) { \
-         vcos_log_error("%d: expected userdata %d, got %d", __LINE__, _u, (int)VCHIQ_GET_SERVICE_USERDATA(service)); VCOS_BKPT; goto error_exit; \
+      else if ((intptr_t)VCHIQ_GET_SERVICE_USERDATA(service) != _u) { \
+         vcos_log_error("%d: expected userdata %d, got %zd", __LINE__, _u, (intptr_t)VCHIQ_GET_SERVICE_USERDATA(service)); VCOS_BKPT; goto error_exit; \
       } \
-      else if ((int)bulk_userdata != _bu) { \
-         vcos_log_error("%d: expected bulk_userdata %d, got %d", __LINE__, _bu, (int)bulk_userdata); VCOS_BKPT; goto error_exit; \
+      else if ((intptr_t)bulk_userdata != _bu) { \
+         vcos_log_error("%d: expected bulk_userdata %d, got %zd", __LINE__, _bu, (intptr_t)bulk_userdata); VCOS_BKPT; goto error_exit; \
       } \
       else \
       { \
